+++ ec2-tui/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 0.2.1
+
+* ESC key backs out of application logs and help screen
+* CTRL+R allows you to resize an EC2 instance from a searchable list of instance types
+* UI now has a "breathing" feel to it by slowly fading between a few colors
+
 ### 0.1.2
 
 * Added regex filter to `ec2-tui`